Replace detach with kill for screen
Detaching from screen still leaves it running in the background which causes people to need to use `sudo` to attach to it again.
Instead if we kill the screen instance with `k` then we can just attach screen again in unprivileged mode.
You also need to confirm the exit by pressing `y`.